+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add endpoints to fetch files as base64 for identity, contacts, companies and bills
 * Add option to remove files for identity, contacts and companies - if the file upload id in the payload is missing, it's ignored, if it's explicitly set to undefined, the file is removed
+* Fix blank email validation for contacts and identities
 
 # 0.4.2